Tidy up the WBE cache (now WBE types) module
Instead of using the expiring cache type as a means
to store worker information just avoid using that type
since we don't support expiry in the first place on
worker information and use a worker container and a
worker object that we can later extend as needed.

Also add on clear methods to the cache type that will be
used when the WBE executor stop occurs. This ensures we
clear out the worker information and any unfinished
requests.

Change-Id: I6a520376eff1e8a6edcef0a59f2d8b9c0eb15752