Provide a deprecated alias for the now removed stop watch class
This allows those who were using it to still continue using it
until 2.0 where it will be removed; this makes it possible for
those users to get off that code in a way that will be easily
do-able (without totally breaking there code-bases, until we do
that in the 2.0 release).

This also removes all internal usage of that stop watch class
so that the library itself will not reference it anymore.

Change-Id: If313d8e7b9bdc8741db2e2e1dfb381aa3260b971