Add deprecated and only alias modules for the moved types
Provide modules for the moved futurist types that only provide
aliases to the old classes/functions and emit a deprecation warning
when imported. This allows for users of those classes/functions to
get off that code in a way that will be easily do-able (without totally
breaking there code-bases, until we do that in the 2.0 release).

Change-Id: If25ac34b437e2a575fe89b98abccbc9a56326d53